STABLE: Put the tail class in another file
